requirements.txt
Pillow>=9.0.0
pytest>=7.0
